Remove aggregationConverter and push its cleanup to aggregationConverters instead
